requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
Brlapi==0.6.6
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.7
command-not-found==0.3
configparser==3.7.4
cryptography==2.1.4
cupshelpers==1.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info==0.18
entrypoints==0.3
Flask==1.0.3
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-bigquery==1.15.0
google-cloud-core==1.0.2
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
httplib2==0.9.2
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
notebook==5.7.8
oauth==1.0.1
olefile==0.45.1
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyRFC3339==1.0
pyrsistent==0.15.2
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.2
qtconsole==4.5.1
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.12.0
system-service==0.3
systemd-python==234
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
xkit==0.0.0
zope.interface==4.3.2
